Fix change the type of the argument to long to handle long integers

+++ _putint.c
@@ -6,7 +6,7 @@
  *
  * Return: number of characters printed
  */
-int _putint(int num)
+int _putint(long int num)
 {
 	int count;
